Refactor NextAuth configuration and update import paths
