+++ src/lib/auth.ts
@@ -1,0 +1,82 @@
+import { NextAuthOptions } from 'next-auth'
+import CredentialsProvider from 'next-auth/providers/credentials'
+import connectDB from '@/lib/mongodb'
+import User from '@/models/User'
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    CredentialsProvider({
+      name: 'credentials',
+      credentials: {
+        email: { label: 'Email', type: 'email' },
+        password: { label: 'Password', type: 'password' }
+      },
+      async authorize(credentials) {
+        if (!credentials?.email || !credentials?.password) {
+          throw new Error('Email and password are required')
+        }
+
+        try {
+          await connectDB()
+          
+          const user = await User.findOne({ 
+            email: credentials.email,
+            isActive: true 
+          }).select('+password')
+
+          if (!user) {
+            throw new Error('Invalid credentials')
+          }
+
+          const isPasswordValid = await user.comparePassword(credentials.password)
+          
+          if (!isPasswordValid) {
+            throw new Error('Invalid credentials')
+          }
+
+          return {
+            id: user._id.toString(),
+            email: user.email,
+            firstName: user.firstName,
+            lastName: user.lastName,
+            role: user.role,
+          }
+        } catch (error) {
+          console.error('Auth error:', error)
+          if (error instanceof Error && error.message.includes('Invalid credentials')) {
+            throw error
+          }
+          throw new Error('Authentication failed')
+        }
+      }
+    })
+  ],
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user) {
+        token.role = user.role
+        token.firstName = user.firstName
+        token.lastName = user.lastName
+      }
+      return token
+    },
+    async session({ session, token }) {
+      if (token) {
+        session.user.id = token.sub!
+        session.user.role = token.role as string
+        session.user.firstName = token.firstName as string
+        session.user.lastName = token.lastName as string
+      }
+      return session
+    }
+  },
+  pages: {
+    signIn: '/login',
+    error: '/login',
+  },
+  session: {
+    strategy: 'jwt',
+    maxAge: 30 * 24 * 60 * 60, // 30 days
+  },
+  secret: process.env.NEXTAUTH_SECRET,
+}
